Move variance calculation service behind facade
The public variance-tracking module remains the import facade while the calculation service moves into the focused variance-tracking service directory.

Constraint: Keep BaselineService AlertManagementService VarianceCalculationService VarianceTrackingService and varianceTrackingService exported from the facade

Confidence: high

Scope-risk: moderate

Tested: npm test -- tests/unit/services/variance-tracking.test.ts tests/unit/services/variance-tracking/variance-diff.test.ts tests/unit/services/variance-tracking/distribution-variance.test.ts tests/unit/services/variance-tracking/company-variance.test.ts

Tested: npm test -- tests/unit/api/variance-tracking-api.test.ts tests/unit/services/variance-alert-automation.test.ts

Tested: npm run check
